backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
pymongo==4.8.0
spotipy==2.23.0
python-dotenv==1.0.0